Initial commit for refactor that moves current transform information out of the VikingViewerControl and into Volume and Section ViewModels where it belongs.
Slight optimization of Section<->Mosaic transform for arrays of points

+++ Clients/WebAnnotationModel/Objects/LocationObj.cs
@@ -298,7 +298,7 @@ namespace WebAnnotationModel
         /// <summary>
         /// Record the hashcode of the volume transform used to map the location. 
         /// </summary>
-        public int? VolumeTransformID = new int?();
+        public long? VolumeTransformID = new int?();
 
         /// <summary>
         /// Return true if the location's volume position has not yet been mapped by this Viking client
